+++ src/main/java/com/Foodie/App/webservice/service/UsersService.java
@@ -1,12 +1,16 @@
 package com.Foodie.App.webservice.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.Foodie.App.webservice.entity.Users;
+import com.Foodie.App.webservice.exception.BadRequestException;
+import com.Foodie.App.webservice.exception.NotFoundException;
 import com.Foodie.App.webservice.repository.UsersRepository;
+import com.Foodie.App.webservice.util.StringUtil;
 
 @Service
 public class UsersService {
@@ -20,27 +24,47 @@ public class UsersService {
 	}
 	
 	//Get one User by UserId
-	public Users getUser(int UserId) {
-		return usersRepository.findById(UserId).get();
+	public Optional<Users> getUser(int UserId) {
+		//return usersRepository.findById(UserId);
+		
+		Optional<Users> user = usersRepository.findById(UserId);
+		if(user.isPresent() && StringUtil.isNotNull(user))
+			return user;
+		else 
+			throw new NotFoundException("Users with provided userId not found");
 	}
 	
 	//Add User
 	public Users addUser(Users user) {
+		//return usersRepository.save(user);
+		
+		if(usersRepository.existsByEmail(user.getEmail()))
+			throw new BadRequestException("User with this email address already exists.");
+
 		return usersRepository.save(user);
 	}
 	
 	//Update User
 	public Users updateUser(Users user) {
+		
+		if(user.getUserId() <=0 )
+			throw new BadRequestException("userId cannot be null or empty.");
+
 		if(usersRepository.existsById(user.getUserId()))
 			return usersRepository.save(user);
-		else
-			return null;
+		else 
+			throw new NotFoundException("The user does not exist with provided userId.");
 				
 	}
 	
 	//Delete User
 	public void deleteUser(int userId) {
-		usersRepository.deleteById(userId);
 				
+		if(userId <=0 )
+			throw new BadRequestException("userId cannot be null or empty.");
+		if(usersRepository.existsById(userId))
+			usersRepository.deleteById(userId);
+		else 
+			throw new NotFoundException("The user does not exist with provided userId.");
 	}
 }
